move sparql proxy cleaning methods into separate SparqlProxyHeaderCleaner class
